refactor: extract pipe-finding logic from run and package
